try 1

+++ utils/str_utils.c
@@ -6,7 +6,7 @@
 /*   By: minylee <minylee@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/02 22:24:34 by minylee           #+#    #+#             */
-/*   Updated: 2023/03/10 15:46:43 by minylee          ###   ########.fr       */
+/*   Updated: 2023/03/10 15:51:36 by minylee          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ size_t	ft_strlen(const char *s)
 
 size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (size != 0)
